chore: update paths for shadcn + toast to sonner
