+++ app/product/[handle]/page.tsx
@@ -1,6 +1,7 @@
 import type { Metadata } from 'next';
 import { notFound } from 'next/navigation';
 
+import { Product } from '@/lib/vendure/types';
 import { GridTileImage } from 'components/grid/tile';
 import Footer from 'components/layout/footer';
 import { Gallery } from 'components/product/gallery';
@@ -9,7 +10,6 @@ import { ProductDescription } from 'components/product/product-description';
 import { getActiveChannel, getProduct } from 'lib/vendure';
 import Link from 'next/link';
 import { Suspense } from 'react';
-import { Product } from '../../../lib/vendure/types';
 
 export async function generateMetadata(props: {
   params: Promise<{ handle: string }>;
