+++ components/cart/close-cart.tsx
@@ -1,6 +1,5 @@
-import { cn } from '@/ui-components/lib/utils';
+import { cn } from '@/lib/utils';
 import { X } from 'lucide-react';
-
 export default function CloseCart({ className }: { className?: string }) {
   return (
     <div className="relative flex h-11 w-11 items-center justify-center rounded-md border border-neutral-200 text-black transition-colors dark:border-neutral-700 dark:text-white">
