+++ components/cart/edit-item-quantity-button.tsx
@@ -1,18 +1,17 @@
 'use client';
 
-import { cn } from '@/ui-components/lib/utils';
+import { cn } from '@/lib/utils';
 import { updateItemQuantity } from 'components/cart/actions';
 import type { OrderLine } from 'lib/vendure/types';
 import { Minus, Plus } from 'lucide-react';
 import { useActionState } from 'react';
-
 function SubmitButton({ type }: { type: 'plus' | 'minus' }) {
   return (
     <button
       type="submit"
       aria-label={type === 'plus' ? 'Increase item quantity' : 'Reduce item quantity'}
       className={cn(
-        'ease flex h-full min-w-[36px] max-w-[36px] flex-none items-center justify-center rounded-full p-2 transition-all duration-200 hover:border-neutral-800 hover:opacity-80',
+        'ease flex h-full max-w-[36px] min-w-[36px] flex-none items-center justify-center rounded-full p-2 transition-all duration-200 hover:border-neutral-800 hover:opacity-80',
         {
           'ml-auto': type === 'minus'
         }
