+++ components/cart/open-cart.tsx
@@ -1,6 +1,5 @@
-import { cn } from '@/ui-components/lib/utils';
+import { cn } from '@/lib/utils';
 import { ShoppingCart } from 'lucide-react';
-
 export default function OpenCart({
   className,
   quantity
@@ -13,7 +12,7 @@ export default function OpenCart({
       <ShoppingCart className={cn('h-4 transition-all ease-in-out hover:scale-110', className)} />
 
       {quantity ? (
-        <div className="absolute right-0 top-0 -mr-2 -mt-2 h-4 w-4 rounded bg-blue-600 text-[11px] font-medium text-white">
+        <div className="absolute top-0 right-0 -mt-2 -mr-2 h-4 w-4 rounded bg-blue-600 text-[11px] font-medium text-white">
           {quantity}
         </div>
       ) : null}
