+++ components/grid/index.tsx
@@ -1,5 +1,4 @@
-import { cn } from '@/ui-components/lib/utils';
-
+import { cn } from '@/lib/utils';
 function Grid(props: React.ComponentProps<'ul'>) {
   return (
     <ul {...props} className={cn('grid grid-flow-row gap-4', props.className)}>
