+++ components/icons/logo.tsx
@@ -1,5 +1,4 @@
-import { cn } from '@/ui-components/lib/utils';
-
+import { cn } from '@/lib/utils';
 export default function LogoIcon(props: React.ComponentProps<'svg'>) {
   return (
     <svg
