+++ components/layout/footer-menu.tsx
@@ -1,11 +1,10 @@
 'use client';
 
-import { cn } from '@/ui-components/lib/utils';
+import { cn } from '@/lib/utils';
 import { CollectionsQuery } from 'lib/vendure/types';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import { useEffect, useState } from 'react';
-
 export function FooterMenuItem({ item }: { item: CollectionsQuery['collections']['items'][0] }) {
   const pathname = usePathname();
   const [active, setActive] = useState(pathname === item.slug);
@@ -19,7 +18,7 @@ export function FooterMenuItem({ item }: { item: CollectionsQuery['collections']
       <Link
         href={item.slug}
         className={cn(
-          'block p-2 text-lg underline-offset-4 hover:text-black hover:underline dark:hover:text-neutral-300 md:inline-block md:text-sm',
+          'block p-2 text-lg underline-offset-4 hover:text-black hover:underline md:inline-block md:text-sm dark:hover:text-neutral-300',
           {
             'text-black dark:text-neutral-300': active
           }
