+++ components/logo-square.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/ui-components/lib/utils';
+import { cn } from '@/lib/utils';
 import LogoIcon from './icons/logo';
 
 export default function LogoSquare({ size }: { size?: 'sm' | undefined }) {
