+++ lib/utils.ts
@@ -1,5 +1,11 @@
+import { clsx, type ClassValue } from 'clsx';
 import { ReadonlyURLSearchParams } from 'next/navigation';
+import { twMerge } from 'tailwind-merge';
 import { SearchResultFragment } from './vendure/types';
+
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
 
 export const createUrl = (pathname: string, params: URLSearchParams | ReadonlyURLSearchParams) => {
   const paramsString = params.toString();
