+++ package.json
@@ -28,6 +28,7 @@
     "graphql-tag": "^2.12.6",
     "lucide-react": "^0.473.0",
     "next": "15.3.0-canary.13",
+    "next-themes": "^0.4.6",
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "react-hook-form": "^7.54.2",
@@ -42,13 +43,13 @@
     "@graphql-codegen/typescript-operations": "^4.4.0",
     "@graphql-typed-document-node/core": "^3.2.0",
     "@tailwindcss/container-queries": "^0.1.1",
-    "autoprefixer": "^10.4.20",
-    "env-cmd": "^10.1.0",
     "@tailwindcss/postcss": "^4.0.14",
     "@tailwindcss/typography": "^0.5.16",
     "@types/node": "22.13.10",
     "@types/react": "19.0.12",
     "@types/react-dom": "19.0.4",
+    "autoprefixer": "^10.4.20",
+    "env-cmd": "^10.1.0",
     "postcss": "^8.5.3",
     "prettier": "3.5.3",
     "prettier-plugin-tailwindcss": "^0.6.11",
